8273946: Move clearQuad method to BaseShaderGraphics superclass
