+++ modules/javafx.graphics/src/main/java/com/sun/prism/d3d/D3DGraphics.java
@@ -26,17 +26,15 @@
 package com.sun.prism.d3d;
 
 import com.sun.javafx.geom.transform.Affine3D;
-import com.sun.prism.CompositeMode;
 import com.sun.prism.Graphics;
 import com.sun.prism.RenderTarget;
 import com.sun.prism.impl.PrismSettings;
 import com.sun.prism.impl.ps.BaseShaderGraphics;
 import com.sun.prism.paint.Color;
-import com.sun.prism.paint.Paint;
 
 class D3DGraphics extends BaseShaderGraphics implements D3DContextSource {
 
-    private D3DContext context;
+    private final D3DContext context;
 
     private D3DGraphics(D3DContext context, RenderTarget target) {
         super(context, target);
@@ -63,24 +61,6 @@ class D3DGraphics extends BaseShaderGraphics implements D3DContextSource {
         }
 
         return new D3DGraphics(context, target);
-    }
-
-    @Override
-    public void clearQuad(float x1, float y1, float x2, float y2) {
-        // note that unlike clear(), this method does not currently
-        // attempt to clear the depth buffer...
-        context.setRenderTarget(this);
-        context.flushVertexBuffer();
-        // set the blend mode to CLEAR and any regular Color as paint
-        CompositeMode oldMode = getCompositeMode();
-        setCompositeMode(CompositeMode.CLEAR);
-        Paint oldPaint = getPaint();
-        setPaint(Color.BLACK); // any color will do...
-        fillQuad(x1, y1, x2, y2);
-        context.flushVertexBuffer();
-        // restore prior paint and blend mode
-        setPaint(oldPaint);
-        setCompositeMode(oldMode);
     }
 
     @Override
